+++ lib/readymade/background_bang_job.rb
@@ -3,9 +3,7 @@
 require 'active_job' unless defined?(::ActiveJob)
 
 module Readymade
-  class BackgroundBangJob < ::ActiveJob::Base
-    queue_as { self.arguments[0].fetch(:queue_as, :default) }
-
+  class BackgroundBangJob < ::Readymade::BackgroundJob
     def perform(**args)
       args.delete(:class_name).to_s.constantize.send(:call!, **args)
     end