+++ spec/readymade/background_job_spec.rb
@@ -16,7 +16,7 @@ RSpec.describe Readymade::BackgroundJob do
   describe '#perform_later' do
     context 'without queue name' do
       it 'creates instance variables from arguments' do
-        allow_any_instance_of(DummyBang).to receive(:call_async).with(**args)
+        allow_any_instance_of(Dummy).to receive(:call_async).with(**args)
 
         res = described_class.perform_later(**args.merge!(class_name: dummy_class.name))
 
@@ -27,12 +27,24 @@ RSpec.describe Readymade::BackgroundJob do
 
     context 'with queue name' do
       it 'creates instance variables from arguments' do
-        allow_any_instance_of(DummyBang).to receive(:call_async).with(**args.merge!(queue_as: :test))
+        allow_any_instance_of(Dummy).to receive(:call_async).with(**args.merge!(queue_as: :test))
 
         res = described_class.perform_later(**args.merge!(class_name: dummy_class.name))
 
         expect(res.job_id.size).to eq(36)
         expect(res.queue_name).to eq('test')
+      end
+
+      context 'when as job options provided' do
+        it 'uses the provided job options' do
+          job_options = { queue_as: :test }
+          allow_any_instance_of(Dummy).to receive(:call_async).with(**args.merge!(job_options: job_options))
+
+          res = described_class.perform_later(**args.merge!(class_name: dummy_class.name, job_options: job_options))
+          expect(res.job_id.size).to eq(36)
+          expect(res.queue_name).to eq('test')
+          expect(res.arguments.first[:job_options]).to eq(job_options)
+        end
       end
     end
   end